Separate precompiles to individual crates
