+++ Cargo.toml
@@ -2,6 +2,11 @@
 members = [
 	"frame/ethereum",
 	"frame/evm",
+	"frame/evm/precompile/simple",
+	"frame/evm/precompile/modexp",
+	"frame/evm/precompile/ed25519",
+	"frame/evm/precompile/bn128",
+	"frame/evm/precompile/blake2",
 	"client/consensus",
 	"client/rpc-core",
 	"client/rpc",
