+++ frame/evm/Cargo.toml
@@ -30,21 +30,9 @@ evm = { version = "0.19.0", default-features = false, features = ["with-codec"] 
 evm-runtime = { version = "0.19.0", default-features = false }
 evm-gasometer = { version = "0.19.0", default-features = false }
 sha3 = { version = "0.8", default-features = false }
-impl-trait-for-tuples = "0.1"
-ripemd160 = { version = "0.9", default-features = false }
-
-num = { version = "0.3", features = ["alloc"], default-features = false, optional = true }
-ethereum-types = { version = "0.9.2", default-features = false, optional = true }
-bn = { package = "substrate-bn", version = "0.5", default-features = false, optional = true }
-ed25519-dalek = { version = "1.0.0", features = ["alloc", "u64_backend"], default-features = false, optional = true }
-
 
 [features]
-default = ["std", "modexp", "ed25519", "bn128", "blake2f"]
-blake2f = []
-modexp = ["num"]
-ed25519 = ["ed25519-dalek"]
-bn128 = ["bn", "ethereum-types"]
+default = ["std"]
 std = [
 	"serde",
 	"codec/std",
@@ -64,5 +52,4 @@ std = [
 	"evm-runtime/std",
 	"evm-gasometer/std",
 	"pallet-timestamp/std",
-	"ripemd160/std",
 ]
