+++ frame/evm/precompile/ed25519/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "pallet-evm-precompile-ed25519"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+description = "ED25519 precompiles for EVM pallet."
+
+[dependencies]
+sp-core = { version = "2.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
+sp-io = { version = "2.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
+fp-evm = { version = "0.8.0", default-features = false, path = "../../../../primitives/evm" }
+evm = { version = "0.19.0", default-features = false, features = ["with-codec"] }
+ed25519-dalek = { version = "1.0.0", features = ["alloc", "u64_backend"], default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"sp-core/std",
+	"sp-io/std",
+	"fp-evm/std",
+	"evm/std",
+	"ed25519-dalek/std",
+]
