+++ primitives/evm/Cargo.toml
@@ -19,6 +19,7 @@ sp-std = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git
 serde = { version = "1.0.101", optional = true, features = ["derive"] }
 codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false }
 evm = { version = "0.19.0", default-features = false, features = ["with-codec"] }
+impl-trait-for-tuples = "0.1"
 
 [features]
 default = ["std"]
