+++ primitives/evm/src/lib.rs
@@ -17,6 +17,8 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+mod precompile;
+
 use codec::{Encode, Decode};
 #[cfg(feature = "std")]
 use serde::{Serialize, Deserialize};
@@ -25,6 +27,7 @@ use sp_core::{U256, H160};
 use evm::ExitReason;
 
 pub use evm::backend::{Basic as Account, Log};
+pub use precompile::{Precompile, PrecompileSet, LinearCostPrecompile};
 
 #[derive(Clone, Eq, PartialEq, Encode, Decode, Default)]
 #[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
